+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## 0.17.0-0.2.1 (upcoming)
 
+* Add Azure file CSI driver
 * Fix wait conditions for unmanaged clusters
 
 ## 0.17.0-0.2.0 (2023-07-05)